+++ hutool-log/src/main/java/cn/hutool/log/dialect/slf4j/Slf4jLog.java
@@ -19,10 +19,13 @@ public class Slf4jLog extends AbstractLog {
 	private static final long serialVersionUID = -6843151523380063975L;
 
 	private final transient Logger logger;
+	/** 是否为 LocationAwareLogger ，用于判断是否可以传递FQCN */
+	private final boolean isLocationAwareLogger;
 
 	// ------------------------------------------------------------------------- Constructor
 	public Slf4jLog(Logger logger) {
 		this.logger = logger;
+		this.isLocationAwareLogger = (logger instanceof LocationAwareLogger);
 	}
 
 	public Slf4jLog(Class<?> clazz) {
@@ -47,7 +50,9 @@ public class Slf4jLog extends AbstractLog {
 	@Override
 	public void trace(String fqcn, Throwable t, String format, Object... arguments) {
 		if (isTraceEnabled()) {
-			if (false == locationAwareLog(fqcn, LocationAwareLogger.TRACE_INT, t, format, arguments)) {
+			if(this.isLocationAwareLogger) {
+				locationAwareLog((LocationAwareLogger)this.logger, fqcn, LocationAwareLogger.TRACE_INT, t, format, arguments);
+			} else {
 				logger.trace(StrUtil.format(format, arguments), t);
 			}
 		}
@@ -62,7 +67,9 @@ public class Slf4jLog extends AbstractLog {
 	@Override
 	public void debug(String fqcn, Throwable t, String format, Object... arguments) {
 		if (isDebugEnabled()) {
-			if (false == locationAwareLog(fqcn, LocationAwareLogger.DEBUG_INT, t, format, arguments)) {
+			if(this.isLocationAwareLogger) {
+				locationAwareLog((LocationAwareLogger)this.logger, fqcn, LocationAwareLogger.DEBUG_INT, t, format, arguments);
+			} else {
 				logger.debug(StrUtil.format(format, arguments), t);
 			}
 		}
@@ -77,7 +84,9 @@ public class Slf4jLog extends AbstractLog {
 	@Override
 	public void info(String fqcn, Throwable t, String format, Object... arguments) {
 		if (isInfoEnabled()) {
-			if (false == locationAwareLog(fqcn, LocationAwareLogger.INFO_INT, t, format, arguments)) {
+			if(this.isLocationAwareLogger) {
+				locationAwareLog((LocationAwareLogger)this.logger, fqcn, LocationAwareLogger.INFO_INT, t, format, arguments);
+			} else {
 				logger.info(StrUtil.format(format, arguments), t);
 			}
 		}
@@ -92,7 +101,9 @@ public class Slf4jLog extends AbstractLog {
 	@Override
 	public void warn(String fqcn, Throwable t, String format, Object... arguments) {
 		if (isWarnEnabled()) {
-			if (false == locationAwareLog(fqcn, LocationAwareLogger.WARN_INT, t, format, arguments)) {
+			if(this.isLocationAwareLogger) {
+				locationAwareLog((LocationAwareLogger)this.logger, fqcn, LocationAwareLogger.WARN_INT, t, format, arguments);
+			} else {
 				logger.warn(StrUtil.format(format, arguments), t);
 			}
 		}
@@ -107,7 +118,9 @@ public class Slf4jLog extends AbstractLog {
 	@Override
 	public void error(String fqcn, Throwable t, String format, Object... arguments) {
 		if (isErrorEnabled()) {
-			if (false == locationAwareLog(fqcn, LocationAwareLogger.ERROR_INT, t, format, arguments)) {
+			if(this.isLocationAwareLogger) {
+				locationAwareLog((LocationAwareLogger)this.logger, fqcn, LocationAwareLogger.ERROR_INT, t, format, arguments);
+			} else {
 				logger.error(StrUtil.format(format, arguments), t);
 			}
 		}
@@ -116,27 +129,25 @@ public class Slf4jLog extends AbstractLog {
 	// ------------------------------------------------------------------------- Log
 	@Override
 	public void log(String fqcn, Level level, Throwable t, String format, Object... arguments) {
-		int level_int;
 		switch (level) {
 		case TRACE:
-			level_int = LocationAwareLogger.TRACE_INT;
+			trace(fqcn, t, format, arguments);
 			break;
 		case DEBUG:
-			level_int = LocationAwareLogger.DEBUG_INT;
+			debug(fqcn, t, format, arguments);
 			break;
 		case INFO:
-			level_int = LocationAwareLogger.INFO_INT;
+			info(fqcn, t, format, arguments);
 			break;
 		case WARN:
-			level_int = LocationAwareLogger.WARN_INT;
+			warn(fqcn, t, format, arguments);
 			break;
 		case ERROR:
-			level_int = LocationAwareLogger.ERROR_INT;
+			error(fqcn, t, format, arguments);
 			break;
 		default:
 			throw new Error(StrUtil.format("Can not identify level: {}", level));
 		}
-		this.locationAwareLog(fqcn, level_int, t, format, arguments);
 	}
 
 	// -------------------------------------------------------------------------------------------------- Private method
@@ -144,6 +155,7 @@ public class Slf4jLog extends AbstractLog {
 	 * 打印日志<br>
 	 * 此方法用于兼容底层日志实现，通过传入当前包装类名，以解决打印日志中行号错误问题
 	 * 
+	 * @param logger {@link LocationAwareLogger} 实现
 	 * @param fqcn 完全限定类名(Fully Qualified Class Name)，用于纠正定位错误行号
 	 * @param level_int 日志级别，使用LocationAwareLogger中的常量
 	 * @param t 异常
@@ -151,15 +163,10 @@ public class Slf4jLog extends AbstractLog {
 	 * @param arguments 参数
 	 * @return 是否支持 LocationAwareLogger对象，如果不支持需要日志方法调用被包装类的相应方法
 	 */
-	private boolean locationAwareLog(String fqcn, int level_int, Throwable t, String msgTemplate, Object[] arguments) {
-		if (this.logger instanceof LocationAwareLogger) {
-			// ((LocationAwareLogger)this.logger).log(null, fqcn, level_int, msgTemplate, arguments, t);
-			// 由于slf4j-log4j12中此方法的实现存在bug，故在此拼接参数
-			((LocationAwareLogger) this.logger).log(null, fqcn, level_int, StrUtil.format(msgTemplate, arguments), null, t);
-			return true;
-		} else {
-			return false;
-		}
+	private void locationAwareLog(LocationAwareLogger logger, String fqcn, int level_int, Throwable t, String msgTemplate, Object[] arguments) {
+		// ((LocationAwareLogger)this.logger).log(null, fqcn, level_int, msgTemplate, arguments, t);
+		// 由于slf4j-log4j12中此方法的实现存在bug，故在此拼接参数
+		logger.log(null, fqcn, level_int, StrUtil.format(msgTemplate, arguments), null, t);
 	}
 
 	/**